Extract notification parameters to dedicated modules #4
